Move visitors to own classes and package
They were getting quite large and making the methods hard to follow.
