+++ src/main/java/com/github/firmwehr/gentle/semantic/SemanticAnalyzer.java
@@ -12,7 +12,9 @@ import com.github.firmwehr.gentle.parser.ast.basictype.BooleanType;
 import com.github.firmwehr.gentle.parser.ast.basictype.IdentType;
 import com.github.firmwehr.gentle.parser.ast.basictype.IntType;
 import com.github.firmwehr.gentle.parser.ast.basictype.VoidType;
-import com.github.firmwehr.gentle.parser.ast.expression.BinaryOperator;
+import com.github.firmwehr.gentle.semantic.analysis.MainMethodLookupVisitor;
+import com.github.firmwehr.gentle.semantic.analysis.SideEffectVisitor;
+import com.github.firmwehr.gentle.semantic.analysis.TypecheckVisitor;
 import com.github.firmwehr.gentle.semantic.ast.LocalVariableDeclaration;
 import com.github.firmwehr.gentle.semantic.ast.SClassDeclaration;
 import com.github.firmwehr.gentle.semantic.ast.SField;
@@ -22,21 +24,6 @@ import com.github.firmwehr.gentle.semantic.ast.basictype.SBasicType;
 import com.github.firmwehr.gentle.semantic.ast.basictype.SBooleanType;
 import com.github.firmwehr.gentle.semantic.ast.basictype.SClassType;
 import com.github.firmwehr.gentle.semantic.ast.basictype.SIntType;
-import com.github.firmwehr.gentle.semantic.ast.expression.SArrayAccessExpression;
-import com.github.firmwehr.gentle.semantic.ast.expression.SBinaryOperatorExpression;
-import com.github.firmwehr.gentle.semantic.ast.expression.SExpression;
-import com.github.firmwehr.gentle.semantic.ast.expression.SLocalVariableExpression;
-import com.github.firmwehr.gentle.semantic.ast.expression.SMethodInvocationExpression;
-import com.github.firmwehr.gentle.semantic.ast.expression.SNewArrayExpression;
-import com.github.firmwehr.gentle.semantic.ast.expression.SNewObjectExpression;
-import com.github.firmwehr.gentle.semantic.ast.expression.SSystemOutPrinlnExpression;
-import com.github.firmwehr.gentle.semantic.ast.expression.SSystemOutWriteExpression;
-import com.github.firmwehr.gentle.semantic.ast.expression.SUnaryOperatorExpression;
-import com.github.firmwehr.gentle.semantic.ast.statement.SExpressionStatement;
-import com.github.firmwehr.gentle.semantic.ast.statement.SIfStatement;
-import com.github.firmwehr.gentle.semantic.ast.statement.SReturnStatement;
-import com.github.firmwehr.gentle.semantic.ast.statement.SWhileStatement;
-import com.github.firmwehr.gentle.semantic.ast.type.SExprType;
 import com.github.firmwehr.gentle.semantic.ast.type.SNormalType;
 import com.github.firmwehr.gentle.semantic.ast.type.SVoidType;
 import com.github.firmwehr.gentle.semantic.ast.type.SVoidyType;
@@ -44,7 +31,6 @@ import com.github.firmwehr.gentle.source.Source;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 public class SemanticAnalyzer {
 	private final Source source;
@@ -171,195 +157,33 @@ public class SemanticAnalyzer {
 		}
 	}
 
+	/**
+	 * Checks that the program typechecks, i.e. all assignments and usages are well typed.
+	 * <br>
+	 * Note that this will not reject statements like "null = null", as the types in this statement match. The {@link
+	 * #checkAssignments(Namespace)} method performs such checks-
+	 *
+	 * @param classes the classes to analyze
+	 *
+	 * @throws SemanticException if any type error is detected.
+	 */
 	void checkTypes(Namespace<SClassDeclaration> classes) throws SemanticException {
-		// TODO Implement
-		// TODO Don't forget the String type
-		Visitor<Void> visitor = new Visitor<>() {
-			SMethod currentMethod;
-
-			@Override
-			public Optional<Void> visit(SMethod method) throws SemanticException {
-				this.currentMethod = method;
-				return Visitor.super.visit(method);
-			}
-
-			@Override
-			public Optional<Void> visit(SIfStatement ifStatement) throws SemanticException {
-				assertIsBoolean(ifStatement.condition());
-
-				return Visitor.super.visit(ifStatement);
-			}
-
-			@Override
-			public Optional<Void> visit(SWhileStatement whileStatement) throws SemanticException {
-				assertIsBoolean(whileStatement.condition());
-				return Visitor.super.visit(whileStatement);
-			}
-
-			@Override
-			public Optional<Void> visit(SArrayAccessExpression arrayExpression) throws SemanticException {
-				assertIsInt(arrayExpression.expression());
-				return Visitor.super.visit(arrayExpression);
-			}
-
-			@Override
-			public Optional<Void> visit(SNewArrayExpression newArrayExpression) throws SemanticException {
-				assertIsInt(newArrayExpression.size());
-				return Visitor.super.visit(newArrayExpression);
-			}
-
-			@Override
-			public Optional<Void> visit(SUnaryOperatorExpression unaryOperatorExpression) throws SemanticException {
-				switch (unaryOperatorExpression.operator()) {
-					case LOGICAL_NOT, NEGATION -> assertIsBoolean(unaryOperatorExpression.expression());
-				}
-				return Visitor.super.visit(unaryOperatorExpression);
-			}
-
-			@Override
-			public Optional<Void> visit(SBinaryOperatorExpression binaryOperatorExpression) throws SemanticException {
-				switch (binaryOperatorExpression.operator()) {
-					case ASSIGN -> {
-						SExprType rhsType = binaryOperatorExpression.rhs().type();
-						SExprType lhsType = binaryOperatorExpression.lhs().type();
-						if (!rhsType.isAssignableTo(lhsType)) {
-							throw new SemanticException(source, null, "Assignment of incompatible type");
-						}
-					}
-					case EQUAL, NOT_EQUAL -> {
-						SExprType rhsType = binaryOperatorExpression.rhs().type();
-						SExprType lhsType = binaryOperatorExpression.lhs().type();
-
-						if (!lhsType.isAssignableTo(rhsType) && !rhsType.isAssignableTo(lhsType)) {
-							throw new SemanticException(source, null, "Incompatible types in comparison");
-						}
-					}
-					case LOGICAL_OR, LOGICAL_AND -> {
-						assertIsBoolean(binaryOperatorExpression.lhs());
-						assertIsBoolean(binaryOperatorExpression.rhs());
-					}
-					case LESS_THAN, LESS_OR_EQUAL, GREATER_THAN, GREATER_OR_EQUAL, ADD, SUBTRACT, MULTIPLY, DIVIDE,
-						MODULO -> {
-						assertIsInt(binaryOperatorExpression.lhs());
-						assertIsInt(binaryOperatorExpression.rhs());
-					}
-				}
-				return Visitor.super.visit(binaryOperatorExpression);
-			}
-
-			@Override
-			public Optional<Void> visit(SSystemOutWriteExpression systemOutWriteExpression) throws SemanticException {
-				assertIsInt(systemOutWriteExpression.argument());
-				return Visitor.super.visit(systemOutWriteExpression);
-			}
-
-			@Override
-			public Optional<Void> visit(SSystemOutPrinlnExpression systemOutPrinlnExpression) throws SemanticException {
-				assertIsInt(systemOutPrinlnExpression.argument());
-				return Visitor.super.visit(systemOutPrinlnExpression);
-			}
-
-			@Override
-			public Optional<Void> visit(SMethodInvocationExpression methodInvocationExpression)
-				throws SemanticException {
-
-				SMethod target = methodInvocationExpression.method();
-				List<LocalVariableDeclaration> parameters = target.parameters();
-				List<SExpression> arguments = methodInvocationExpression.arguments();
-
-				if (parameters.size() != arguments.size()) {
-					throw new SemanticException(source, null, "Received wrong number of arguments");
-				}
-
-				for (int i = 0; i < parameters.size(); i++) {
-					if (!arguments.get(i).type().isAssignableTo(parameters.get(i).getType())) {
-						throw new SemanticException(source, null, "Mismatched types at index " + i);
-					}
-				}
-
-				return Visitor.super.visit(methodInvocationExpression);
-			}
-
-			@Override
-			public Optional<Void> visit(SReturnStatement returnStatement) throws SemanticException {
-				if (currentMethod == null) {
-					throw new IllegalStateException("Return outside of method");
-				}
-				Optional<SExpression> returnValue = returnStatement.returnValue();
-
-				if (returnValue.isEmpty()) {
-					if (returnStatement.returnValue().isPresent()) {
-						throw new SemanticException(source, null, "Void method must not return anything");
-					}
-					return Visitor.super.visit(returnStatement);
-				}
-
-				if (!returnValue.get().type().isAssignableTo(currentMethod.returnType().asExprType())) {
-					throw new SemanticException(source, null, "Not assignable to return type");
-				}
-
-				return Visitor.super.visit(returnStatement);
-			}
-
-			@Override
-			public Optional<Void> visit(SNewObjectExpression newObjectExpression) throws SemanticException {
-				Optional<SNormalType> normalType = newObjectExpression.type().asNormalType();
-
-				if (normalType.isPresent() && normalType.get().arrayLevel() == 0) {
-					if (normalType.get().basicType().asStringType().isPresent()) {
-						throw new SemanticException(source, null, "Can not create instances of String");
-					}
-				}
-
-				return Visitor.super.visit(newObjectExpression);
-			}
-
-			private void assertIsBoolean(SExpression expression) throws SemanticException {
-				Optional<SNormalType> normalType = expression.type().asNormalType();
-
-				if (normalType.isEmpty() || normalType.get().arrayLevel() != 0) {
-					throw new SemanticException(source, null, "Condition must be a boolean");
-				}
-
-				if (normalType.get().basicType().asBooleanType().isEmpty()) {
-					throw new SemanticException(source, null, "Condition must be a boolean");
-				}
-			}
-
-			private void assertIsInt(SExpression expression) throws SemanticException {
-				Optional<SNormalType> normalType = expression.type().asNormalType();
-
-				if (normalType.isEmpty() || normalType.get().arrayLevel() != 0) {
-					throw new SemanticException(source, null, "Expression must be an integer");
-				}
-
-				if (normalType.get().basicType().asIntType().isEmpty()) {
-					throw new SemanticException(source, null, "Expression must be an integer");
-				}
-			}
-
-		};
+		Visitor<Void> visitor = new TypecheckVisitor(source);
 
 		for (SClassDeclaration declaration : classes.getAll()) {
 			visitor.visit(declaration);
 		}
 	}
 
+	/**
+	 * Checks that all expressions have side effects.
+	 *
+	 * @param classes the classes to analyze
+	 *
+	 * @throws SemanticException if any statement does not have a side effect
+	 */
 	void checkSideEffects(Namespace<SClassDeclaration> classes) throws SemanticException {
-		// TODO Check if all ExpressionStatements have side effects
-
-		Visitor<Void> visitor = new Visitor<>() {
-			@Override
-			public Optional<Void> visit(SExpressionStatement expressionStatement) throws SemanticException {
-				SExpression expression = expressionStatement.expression();
-				return switch (expression) {
-					case SMethodInvocationExpression ignored -> Optional.empty();
-					case SBinaryOperatorExpression op && op.operator() == BinaryOperator.ASSIGN -> Optional.empty();
-					default -> throw new SemanticException(source, null,
-						"Expression statement must habe side " + "effects");
-				};
-			}
-		};
+		Visitor<Void> visitor = new SideEffectVisitor(source);
 
 		for (SClassDeclaration declaration : classes.getAll()) {
 			visitor.visit(declaration);
@@ -374,57 +198,31 @@ public class SemanticAnalyzer {
 		// TODO Implement
 	}
 
+	/**
+	 * Finds the unique main method for all given classes, failing if it does not exist, multiple exist or the main
+	 * method semantics are not respected:
+	 * <ul>
+	 *     <li>There is only one static method</li>
+	 *     <li>The static method is named main</li>
+	 *     <li>The static method has a {@code String[]} parameter</li>
+	 *     <li>The static method returns void</li>
+	 *     <li>The parameter of the static method is not used in the method</li>
+	 * </ul>
+	 *
+	 * @param classes the classes to analyze
+	 *
+	 * @return the found, globally unique, main method
+	 *
+	 * @throws SemanticException if any of the above conditions is violated
+	 */
 	SMethod findMainMethod(Namespace<SClassDeclaration> classes) throws SemanticException {
-		// TODO Implement
-		// TODO Check main method semantics
+		var visitor = new MainMethodLookupVisitor(source);
 
-		Visitor<SMethod> visitor = new Visitor<>() {
-			private SMethod foundMainMethod;
-			private LocalVariableDeclaration mainMethodParameter;
-
-			@Override
-			public Optional<SMethod> visit(SMethod method) throws SemanticException {
-				if (!method.isStatic()) {
-					return Optional.empty();
-				}
-				if (this.foundMainMethod != null) {
-					throw new SemanticException(source, null, "Found second main method!");
-				}
-				if (!method.name().ident().equals("main")) {
-					throw new SemanticException(source, null, "Only 'main' is allowed for static method names");
-				}
-				if (method.parameters().size() != 1) {
-					throw new IllegalArgumentException("The main method must have exactly one parameter");
-				}
-				LocalVariableDeclaration parameter = method.parameters().get(0);
-
-				if (parameter.getType().arrayLevel() != 1) {
-					throw new SemanticException(source, null, "The main method must have a String[] parameter");
-				}
-				if (parameter.getType().basicType().asStringType().isEmpty()) {
-					throw new SemanticException(source, null, "The main method must have a String[] parameter");
-				}
-
-				this.foundMainMethod = method;
-				this.mainMethodParameter = parameter;
-
-				Visitor.super.visit(method);
-				return Optional.of(method);
-			}
-
-			@Override
-			public Optional<SMethod> visit(SLocalVariableExpression localVariableExpression) throws SemanticException {
-				if (localVariableExpression.localVariable() == mainMethodParameter) {
-					throw new SemanticException(source, null, "Usage of main method parameter is forbidden");
-				}
-				return Visitor.super.visit(localVariableExpression);
-			}
-		};
-
-		Optional<SMethod> main = Optional.empty();
 		for (SClassDeclaration declaration : classes.getAll()) {
-			main = visitor.visit(declaration);
+			visitor.visit(declaration);
 		}
-		return main.orElseThrow(() -> new SemanticException(source, null, "Did not find a main method :/"));
+
+		return visitor.getFoundMainMethod();
 	}
+
 }
